Separate printer and labels in configuration file
This lets the server known of a 'database' of common label sizes,
particular those we have. At the moment the main goal is data modelling
as we'll surely move to an actual database if interaction with inventory
is necessary for ordering and keeping track of available counts.

The new data model fits better a workflow where a printer's labels are
changed on the fly, after which the configuration file can be reloaded.