Relax SVG rendering offset to warning
We don't know that the image is within bounds by construction. Sure, we
scale it down to fit in theory but floating point inexactness and the
algorithm choice of resvg may simply change that. Since there is no
influence on this we should not crash the label server's physical
printer task over this but rather report it. Would be nice if we
reported it structurally to that task but that's for tomorrow.